+++ components/profile/source/com/company/itos/profile/person/dao/PersonListDAO.java
@@ -33,7 +33,7 @@ public class PersonListDAO {
 			while (resultSet.next()) {
 
 				PersonDetail personDetail = new PersonDetail();
-				personDetail.setPersonID(resultSet.getInt("personID"));
+				personDetail.setPersonID(resultSet.getLong("personID"));
 				personDetail.setRefrenceNumber(resultSet.getInt("REFRENCENUMBER"));
 				personDetail.setTitle(resultSet.getString("title"));
 				personDetail.setFirstName(resultSet.getString("FIRSTNAME"));
